tests/wpt/tests/css/css-contain/content-visibility/content-visibility-vs-scrollIntoView-003.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>CSS Content Visibility: auto + overflow clip + scrollIntoView, ContentVisibilityAutoStateChange fires twice</title>
<link rel="author" title="Cathie Chen" href="mailto:cathiechen@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-contain/#content-visibility">
<meta name="assert"
    content="If content-visibility: auto element is not relevent to user after calling scrollIntoView of its descendant, contentvisibilityautostatechange twice">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
    .auto {
        content-visibility: auto;
        contain-intrinsic-size: auto 1px auto 10000px;
    }

    .child {
        height: 40000px;
        position: relative;
    }

    #target {
        position: absolute;
        bottom: 0;
    }

    .before_target {
        height: 40000px;
    }

    #overflow_clip {
        overflow: clip;
        height: 20000px;
    }
</style>

<div id=e1 class="auto before_target"></div>
<div id=e2 class="auto before_target"></div>
<div id=e3 class="auto">
    <div id="overflow_clip">
        <div class=child>
            <div id=target>PASS</div>
        </div>
    </div>
</div>
<div id=e4 class=auto>
    <div class=child></div>
</div>

<script>
promise_test(t => new Promise(async (resolve, reject) => {
    await new Promise((waited, _) => {
        requestAnimationFrame(() => requestAnimationFrame(waited));
    });

    function waitForEvent() {
        return new Promise(resolve => e3.addEventListener('contentvisibilityautostatechange', resolve));
    }

    var eventCounter = 0;
    function eventHandler(e) {
        eventCounter++;
        if (eventCounter == 1) {
            assert_equals(e.skipped, false, "the first event should be generated by visible");
        } else if (eventCounter == 2) {
            assert_equals(e.skipped, true, "the second event should be generated by hidden");
        }
    }

    e3.addEventListener("contentvisibilityautostatechange", eventHandler);
    target.scrollIntoView();
    await waitForEvent();
    await waitForEvent();
    requestAnimationFrame(() => requestAnimationFrame(() => {
        assert_equals(eventCounter, 2, "There should be two contentvisibilityautostatechange events.");
        resolve();
    }));
}), "ContentVisibilityAutoStateChange fires twice when `scrollIntoView` a descendant of `content-visibility:auto` which is hidden after scrolling");
</script>
